Regenerate test files
The test cpio files are generated with `--reproducible` which is
equivalent to `--ignore-devno --ignore-dirnlink --renumber-inodes`.
`--ignore-dirnlink` stores 2 in the nlink field of each directory
archive member, instead of the actual number of links.

The current test files have 3 in the nlink field of some directories. So
regenerate the test files to store 2 in the nlink field.