feat: add --data-align option to the --create mode
Add a `--data-align` option to the `--create` mode which will pad the
cpio metadata to align the file data. This option is useful to reflink
cpio file data on file systems that support reflinks.

**Note**: Using this option will "bend" the cpio newc spec a bit to
inject zeros after the filename to provide data segment alignment. These
zeros are accounted for in the namesize, but some applications may only
expect a single zero-terminator (and 4 byte alignment). GNU cpio and
Linux initramfs handle this fine as long as `PATH_MAX` is not exceeded.