refactor: set current dir in read_cpio_and_extract
Move the `set_current_dir` call further down into the
`read_cpio_and_extract` function. That moves the call closer to the
functions that rely on it.

For performance reasons do not restore the current working directory in
`read_cpio_and_extract`. Rely on the caller to do that if needed.